requirements.txt
git+https://github.com/huggingface/transformers
accelerate
qwen-vl-utils[decord]==0.0.8
torch>=2.0.0
torchvision
pillow